chore(cxx_common): move MemcahcedHashCache to its own file
