+++ kythe/cxx/common/indexing/BUILD
@@ -29,9 +29,11 @@ cc_library(
     name = "caching_output",
     srcs = [
         "KytheCachingOutput.cc",
+        "MemcachedHashCache.cc",
     ],
     hdrs = [
         "KytheCachingOutput.h",
+        "MemcachedHashCache.h",
     ],
     visibility = [PUBLIC_VISIBILITY],
     deps = [
