+++ kythe/cxx/common/indexing/KytheCachingOutput.cc
@@ -25,58 +25,6 @@
 #include "absl/strings/str_format.h"
 
 namespace kythe {
-
-bool MemcachedHashCache::OpenMemcache(const std::string& spec) {
-  if (cache_) {
-    memcached_free(cache_);
-    cache_ = nullptr;
-  }
-  std::string spec_amend = spec;
-  spec_amend.append(" --BINARY-PROTOCOL");
-  cache_ = memcached(spec_amend.c_str(), spec_amend.size());
-  if (cache_ != nullptr) {
-    memcached_return_t remote_version = memcached_version(cache_);
-    return memcached_success(remote_version);
-  }
-  return false;
-}
-
-MemcachedHashCache::~MemcachedHashCache() {
-  if (cache_) {
-    memcached_free(cache_);
-    cache_ = nullptr;
-  }
-}
-
-void MemcachedHashCache::RegisterHash(const Hash& hash) {
-  if (!cache_) {
-    return;
-  }
-  char value = 1;
-  memcached_return_t add_result =
-      memcached_add(cache_, reinterpret_cast<const char*>(hash), kHashSize,
-                    &value, sizeof(value), 0, 0);
-  if (!memcached_success(add_result) && add_result != MEMCACHED_DATA_EXISTS) {
-    absl::FPrintF(stderr, "memcached add failed: %s\n",
-                  memcached_strerror(cache_, add_result));
-  }
-}
-
-bool MemcachedHashCache::SawHash(const Hash& hash) {
-  if (!cache_) {
-    return false;
-  }
-  memcached_return_t ex_result =
-      memcached_exist(cache_, reinterpret_cast<const char*>(hash), kHashSize);
-  if (ex_result == MEMCACHED_SUCCESS) {
-    return true;
-  } else if (ex_result != MEMCACHED_NOTFOUND) {
-    absl::FPrintF(stderr, "memcached exist failed: %s\n",
-                  memcached_strerror(cache_, ex_result));
-  }
-  return false;
-}
-
 std::string FileOutputStream::Stats::ToString() const {
   return absl::StrCat(
       buffers_merged_, " merged ", buffers_split_, " split ", buffers_retired_,
